Create primeNum&coinFlipper
This code also has two purposes. The first part will ask the user for a number. It will then determine if that given number is prime; it will also find its factors and the amount of numbers that are prime which are below the given number. The second portion is a coin flipper, it will ask the user if it wants heads or tails and how many of in a row. It will continuously flip a coin until it gets the requested number of consecutive flips of the requested side.